Started splitting of commands in separate files
