+++ viper/core/ui/cmd/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+# This file is part of Viper - https://github.com/viper-framework/viper
+# See the file 'LICENSE' for copying permission.
